Criada classe para gerar números random, mudañça nas classes de ordenação de int para double, assim como na classe Util.

+++ src/sort/BubbleSort.java
@@ -12,7 +12,7 @@ package sort;
 import util.Util;
 
 public class BubbleSort {
-	public static void sort(int[] v) {
+	public static void sort(double[] v) {
 		int i, j, n = v.length;
 		for (i = 0; i < n; i++) {
 			for (j = i + 1; j < n; j++) {
